fix "CreateMessageBox must be called on the UI thread"
If an error occurs when saving the project options on deploy,
Errors.ShowCritical is invoked from an async worker thread,
which results in this perplexing message.

By using BeginInvoke, we can get rid of the requirement
on the callers of CreateMessageBox to be on the UI thread,
which should prevent this from showing up ever again.