optimize text output parsing
The previous approach is very inefficient, particularly when the
output file is located locally. First, the entire file is loaded in
memory; then, it is parsed line-by-line using async (!) StreamReader
API. Even in the Debug Server code, which streams the file, async
calls are useless because FileStream performs blocking IO unless
useAsync: true is passed in the constructor (it isn't).

Using sync APIs is just the start. There's an allocation for every
StreamReader.ReadLine() call, plus another allocation for every
string.Replace() call, so we get 2 allocations per value read.
Given that debug output typically contains on the order of 260000
values, this quickly adds up.

I decided to write a hand-rolled line-delimited hex value parser
to avoid these slowdowns. I experimented with using bit-masks to parse
each ASCII digit, but I found that a simple lookup table is faster
(verified with a microbenchmark using BenchmarkDotNet), not to mention
leads to more understandable code.

The end result, as tested on an FX-8120 machine with ~2MB output files:
* 230-270 ms with the old code,
* 22-23 ms with the new version.

give constants meaningful names