allow fallible predefined macros
If there are no breakpoints set, and the action uses the $(RadBreakLine)
macro, we should always get an error, regardless of whether this is the
debug action or e.g. "assemble debug shader".

On the other hand, if the action does not need to know break lines,
there should be no errors.