visualizer: fix stale status string if group index is out of bounds on break
Also refactor confusing handling of GroupIndexChangedEventArgs: it
contaied properties set by the event handler and read back by the
sender. This can be simplied by computing the number of groups inside
GroupIndexSelector, however I do not want to introduce a dependency on
BreakStateData's GetGroupCount because it makes the tests messier. Since
all we need to compute the number of groups based on group size is the
total number of threads in a buffer, we can pass that instead of a
reference to BreakStateData, and GetGroupCount then becomes obsolete.